pyproject.toml
[tool.poetry]
name = "hiwi-stunden"
version = "0.1.0"
description = ""
authors = ["NotAFile <notafile@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.2"
python-ldap = "^3.2.0"
django-auth-ldap = "^2.1.0"
django-mathfilters = "^0.4.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
